Strip whitespace from reindex and updateByQuery scripts
This is a somewhat no-brainer - stripping whitespace avoids whitespace
errors from Elasticsearch and makes it a litle easier on the user (i.e.
the developer).

With that said, this is not a golden egg. In the `reindex()` method,
only the string-ified version is whitespace-stripped. (I didn't know
enough about the incoming `script` argument data format to say whether
we could manipulate this as well.)