Update default module settings documentation
This commit pastes the cbElasticsearch moduleSettings in order to update the documentation on what the actual settings are by default. For example, the `client` default setting documentation was out of date, as well as `defaultIndexShards`, plus a number of settings were missing entirely. (And only noted in the "secondary Elasticsearch cluster" documentation.)